+++ server/routes.js
@@ -2,8 +2,6 @@ const express = require("express");
 const fs = require("fs");
 const util = require("util");
 const dayjs = require("dayjs");
-const winston = require("winston");
-const expressWinston = require("express-winston");
 
 const { getProject, getTeam, getUser, getZine } = require("./api");
 const initWebpack = require("./webpack");
@@ -28,40 +26,6 @@ module.exports = function(external) {
   const ms = dayjs.convert(7, "days", "miliseconds");
   app.use(express.static("public", { index: false }));
   app.use(express.static("build", { index: false, maxAge: ms }));
-
-  // Log all requests to a local file for diagnostics
-  const requestTime = function(req, res, next) {
-    req.requestTime = Date.now();
-    next();
-  };
-  
-  app.use(requestTime)
-  
-  app.use(
-    expressWinston.logger({
-      transports: [
-        new winston.transports.File({
-          name: "requests",
-          filename: "requests.log",
-          maxsize: 2500, // max size of each log file in bytes
-          maxFiles: 1,
-          tailable: true, // should make it so the oldest data will get removed from the file when it exceeds maxsize
-        }),
-      ],
-      format: winston.format.combine(
-        winston.format.colorize(),
-        winston.format.json(),
-      ),
-      meta: false, // logs meta data about the request if true
-      msg: (req, res) => {
-        return `${req.requestTime}: HTTP ${req.method} ${req.url} ${res.statusCode} /
-Response-Time: ${res.responseTime}ms / 
-User-Agent: ${req.headers["user-agent"]} / 
-Cache-Control: ${req.headers["cache-control"]}`;
-      },
-      colorize: false, // Color the text and status code, using the Express/morgan color palette (text: gray, status: default green, 3XX cyan, 4XX yellow, 5XX red).
-    }),
-  );
 
   const readFilePromise = util.promisify(fs.readFile);
   const imageDefault =
